Added surface anisotropy calculation routines
Surface anisotropy exists only where total number of neighbours is less than threshold, ie: number of nearest neighbours. This is ill defined for long range input, and should instead be set explicitly as a different material for surface atoms. If enabled, arrays storing the unit vectors between nearest neighbours are created when the create system arrays are copied to the atoms namespace. Additionally relevant flag variables are set in vinput. A mask is created identifying the surface atoms, and fields and energies are only calculated for surface atoms.