Modified geometry function to accept material selection by geometry.
By default, geometry assumes a single material per layer. However, a similar routine can be used to have multiple materials in a single layer, for example creating nested structures, with the caveat that this overrides the input materials for example in ordered alloys. This is enabled with the flag create:select-material-by-geometry.