Add selection initial and final frame to render
Create two command line options "--frame-start" and "--frame-final"
that, if specified, allow independent choice of the intial and
final snapshot to render with vdc utility. The frame number is the
real index of the spins-*.data that one wants to render: for instance
"--frame-start 20 --frame-final 350" will process snapshots from
spins-00000020.data to spins-00000350.data. Moreover, if the option
is specified the inital and final frame to render in povray are
updated.